+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,3 +1,13 @@
+## Upcoming
+
+🐞 Fixed
+
+- [[#1592]](https://github.com/GetStream/stream-chat-flutter/issues/1592) Fixed broken attachment download on web.
+
+🔄 Changed
+
+- Updated `dio` dependency to `^5.2.0`.
+
 ## 6.2.0
 
 🐞 Fixed